+++ main.py
@@ -1,0 +1,7 @@
+from TermiText import termitext as termi
+
+termi.animation.create("test", 0.5)
+termi.animation.frame("test", "testing1")
+termi.animation.frame("test", "testing2")
+termi.animation.frame("test", "testing4")
+termi.animation.play("test")